finaly app base

+++ app/database/config.js
@@ -13,6 +13,19 @@ export async function openDatabase() {
   }
 }
 
+const dropTables = async () => {
+  const db = await openDatabase();
+  await db.execAsync(`
+    DROP TABLE IF EXISTS users;
+    DROP TABLE IF EXISTS projects;
+    DROP TABLE IF EXISTS images;
+  `);
+}
+
+// dropTables().catch(error => {
+//   console.error('Falha ao deletar tabelas:', error);
+// })
+
 export async function initializeDatabase() {
   const db = await openDatabase();
   await db.execAsync(`
@@ -32,8 +45,10 @@ export async function initializeDatabase() {
     CREATE TABLE IF NOT EXISTS images (
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       project_id INTEGER,
+      user_id INTEGER,
       path TEXT,
-      FOREIGN KEY (project_id) REFERENCES projects(id)
+      FOREIGN KEY (project_id) REFERENCES projects(id),
+      FOREIGN KEY (user_id) REFERENCES users(id)
     );
   `);
 }
@@ -65,14 +80,14 @@ export async function getUserByEmail(email) {
   }
 }
 
-export async function saveImagePaths(projectId, imagePaths) {
+export async function saveImagePaths(projectId, userId, imagePaths) {
   try {
     const db = await openDatabase();
-    await db.runAsync('DELETE FROM images WHERE project_id = ?;', projectId);
+    await db.runAsync('DELETE FROM images WHERE project_id = ? AND user_id = ?;', projectId, userId);
 
     imagePaths.forEach(async path => {
       if (path) {
-        await db.runAsync('INSERT INTO images (project_id, path) VALUES (?, ?);', projectId, path);
+        await db.runAsync('INSERT INTO images (project_id, user_id, path) VALUES (?, ?, ?);', projectId, userId, path);
       }
     });
 
@@ -83,12 +98,22 @@ export async function saveImagePaths(projectId, imagePaths) {
   }
 }
 
-export async function getAllPathImages() {
+export async function getImagesByUserId(userId) {
   try {
     const db = await openDatabase();
-    return await db.getAllAsync('SELECT * FROM images;');
+    return await db.getAllAsync('SELECT * FROM images WHERE user_id = ?;', userId);
   } catch (error) {
-    console.error('Falha ao obter caminhos das imagens:', error);
+    console.error('Falha ao obter caminhos das imagens por usuário:', error);
+    throw error;
+  }
+}
+
+export async function getImagesByProjectId(userId, projectId) {
+  try {
+    const db = await openDatabase();
+    return await db.getAllAsync('SELECT * FROM images WHERE user_id = ? AND project_id = ?;', userId, projectId);
+  } catch (error) {
+    console.error('Falha ao obter caminhos das imagens por projeto:', error);
     throw error;
   }
 }
